refactor: move Slack client functionality to service layer
